move class FeatureContext
